Refactor participant endpoint utils function into separate file
